requirements.txt
attrs==20.2.0
autopep8==1.5.4
cachelib==0.1.1
certifi==2020.6.20
click==7.1.2
docutils==0.16
elasticsearch==7.9.1
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Sessionstore==0.4.5
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
packaging==20.4
pluggy==0.13.1
psutil==5.7.2
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.1
pytest-mock==3.3.1
python-daemon==2.2.4
pytz==2020.1
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==0.16.0
zipp==3.3.1
